# perf: cache json strings into const